extract password validation into libs and add tests
