+++ src/server/lib/validatePasswordField.ts
@@ -1,0 +1,20 @@
+import { FieldError } from '@/shared/model/ClientState';
+import { ChangePasswordErrors } from '@/shared/model/Errors';
+
+export const validatePasswordField = (password: string): Array<FieldError> => {
+  const errors: Array<FieldError> = [];
+
+  if (!password) {
+    errors.push({
+      field: 'password',
+      message: ChangePasswordErrors.PASSWORD_BLANK,
+    });
+  } else if (password.length <= 8 || password.length > 72) {
+    errors.push({
+      field: 'password',
+      message: ChangePasswordErrors.PASSWORD_LENGTH,
+    });
+  }
+
+  return errors;
+};
